More tests working and some more cleaning / refactoring
