+++ test/ly/stealth/xmlavro/SchemaTests.java
@@ -1,13 +1,13 @@
 package ly.stealth.xmlavro;
 
 import org.apache.avro.Schema;
-import org.apache.avro.generic.GenericData;
+import org.json.JSONException;
 import org.junit.Test;
+import org.skyscreamer.jsonassert.JSONAssert;
 
 import java.util.Arrays;
 
 import static junit.framework.Assert.*;
-import static junit.framework.Assert.assertEquals;
 
 public class SchemaTests {
 
@@ -59,5 +59,63 @@ public class SchemaTests {
                 optionalSchema.getTypes()
         );
     }
+
+    @Test
+    public void array() {
+        Schema schema = Converter.createSchema(TestData.array.xsd);
+        Schema.Field valueField = schema.getField("value");
+        assertEquals(Schema.Type.ARRAY, valueField.schema().getType());
+        assertEquals(Schema.Type.STRING, valueField.schema().getElementType().getType());
+    }
+
+    @Test
+    public void SchemaBuilder_validName() {
+        SchemaBuilder builder = new SchemaBuilder();
+
+        assertNull(builder.validName(null));
+        assertEquals("", builder.validName(""));
+
+        assertEquals("a1", builder.validName("$a#1"));
+
+        assertEquals("a_1", builder.validName("a.1"));
+        assertEquals("a_1", builder.validName("a-1"));
+
+        // built-in types
+        assertEquals("string0", builder.validName("string"));
+        assertEquals("record1", builder.validName("record"));
+    }
+
+    @Test
+    public void arrayOfChoiceElements() {
+        Schema schema = Converter.createSchema(TestData.arrayOfChoiceElements.xsd);
+        assertEquals(Schema.Type.ARRAY, schema.getType());
+        final Schema elementType = schema.getElementType();
+        assertEquals(Schema.Type.RECORD, elementType.getType());
+        assertEquals(2, elementType.getFields().size());
+    }
+
+    @Test
+    public void arrayFromComplexTypeChoiceElements() throws JSONException {
+        Schema schema = Converter.createSchema(TestData.arrayFromComplexTypeChoiceElements.xsd);
+        JSONAssert.assertEquals("{" +
+                "    'type': 'array'," +
+                "    'items': {" +
+                "        'type': 'record'," +
+                "        'name': 'AnonType_root'," +
+                "        'fields': [" +
+                "            {" +
+                "                'name': 's'," +
+                "                'type': ['null', 'string']" +
+                "            }," +
+                "            {" +
+                "                'name': 'i'," +
+                "                'type': ['null', 'int']" +
+                "            }" +
+                "        ]" +
+                "    }" +
+                "}", schema.toString(), false);
+    }
+
+
 
 }
